+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-14 21:05:57
-// Commit: docs(core): update readme and API contracts for core
+// Updated at: 2026-05-15 09:23:19
+// Commit: feat(core): add bi-directional event bus handler for core
 
 class coreComponentState {
-  final String commitRef = '1166';
-  final String lastMessage = 'docs(core): update readme and API contracts for core';
-  final DateTime timestamp = DateTime.parse('2026-05-14T21:05:57+05:00');
+  final String commitRef = '1173';
+  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
+  final DateTime timestamp = DateTime.parse('2026-05-15T09:23:19+05:00');
 }